Allow users to edit user message
