+++ app/src/main/kotlin/dev/chungjungsoo/gptmobile/presentation/ui/chat/ChatScreen.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -24,35 +23,25 @@ import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Done
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material.icons.rounded.KeyboardArrowDown
-import androidx.compose.material.icons.rounded.Refresh
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.BottomAppBarDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FabPosition
-import androidx.compose.material3.FilledTonalButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SmallFloatingActionButton
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.TopAppBarScrollBehavior
@@ -78,11 +67,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.DialogProperties
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dev.chungjungsoo.gptmobile.R
@@ -121,11 +108,13 @@ fun ChatScreen(
 
     val chatRoom by chatViewModel.chatRoom.collectAsStateWithLifecycle()
     val isChatTitleDialogOpen by chatViewModel.isChatTitleDialogOpen.collectAsStateWithLifecycle()
+    val isEditQuestionDialogOpen by chatViewModel.isEditQuestionDialogOpen.collectAsStateWithLifecycle()
     val isIdle by chatViewModel.isIdle.collectAsStateWithLifecycle()
     val isLoaded by chatViewModel.isLoaded.collectAsStateWithLifecycle()
     val messages by chatViewModel.messages.collectAsStateWithLifecycle()
     val question by chatViewModel.question.collectAsStateWithLifecycle()
     val appEnabledPlatforms by chatViewModel.enabledPlatformsInApp.collectAsStateWithLifecycle()
+    val editedQuestion by chatViewModel.editedQuestion.collectAsStateWithLifecycle()
 
     val openaiLoadingState by chatViewModel.openaiLoadingState.collectAsStateWithLifecycle()
     val anthropicLoadingState by chatViewModel.anthropicLoadingState.collectAsStateWithLifecycle()
@@ -147,7 +136,7 @@ fun ChatScreen(
     val canUseChat = (chatViewModel.enabledPlatformsInChat.toSet() - appEnabledPlatforms.toSet()).isEmpty()
     val groupedMessages = remember(messages) { groupMessages(messages) }
     val latestMessageIndex = groupedMessages.keys.maxOrNull() ?: 0
-    val chatBubbleScrollStates = rememberSaveable(saver = multiScrollStateSaver) { DefaultHashMap<Int, ScrollState>({ ScrollState(0) }) }
+    val chatBubbleScrollStates = rememberSaveable(saver = multiScrollStateSaver) { DefaultHashMap<Int, ScrollState> { ScrollState(0) } }
     val canEnableAICoreMode = rememberSaveable { checkAICoreAvailability(aiCorePackageInfo, privateComputePackageInfo) }
 
     val scope = rememberCoroutineScope()
@@ -212,7 +201,9 @@ fun ChatScreen(
                             UserChatBubble(
                                 modifier = Modifier.widthIn(max = maximumChatBubbleWidth),
                                 text = groupedMessages[key]!![0].content,
-                                onCopyClick = { clipboardManager.setText(AnnotatedString(groupedMessages[key]!![0].content.trim())) }
+                                isLoading = !isIdle,
+                                onCopyClick = { clipboardManager.setText(AnnotatedString(groupedMessages[key]!![0].content.trim())) },
+                                onEditClick = { chatViewModel.openEditQuestionDialog(groupedMessages[key]!![0]) }
                             )
                         }
                     }
@@ -257,7 +248,9 @@ fun ChatScreen(
                         UserChatBubble(
                             modifier = Modifier.widthIn(max = maximumChatBubbleWidth),
                             text = userMessage.content,
-                            onCopyClick = { clipboardManager.setText(AnnotatedString(userMessage.content.trim())) }
+                            isLoading = true,
+                            onCopyClick = { clipboardManager.setText(AnnotatedString(userMessage.content.trim())) },
+                            onEditClick = { chatViewModel.openEditQuestionDialog(userMessage) }
                         )
                     }
                 }
@@ -315,6 +308,17 @@ fun ChatScreen(
                 onRetryRequest = chatViewModel::generateAIChatTitle,
                 onConfirmRequest = { title -> chatViewModel.updateChatTitle(title) },
                 onDismissRequest = chatViewModel::closeChatTitleDialog
+            )
+        }
+
+        if (isEditQuestionDialogOpen) {
+            ChatQuestionEditDialog(
+                initialQuestion = editedQuestion,
+                onDismissRequest = chatViewModel::closeEditQuestionDialog,
+                onConfirmRequest = { question ->
+                    chatViewModel.editQuestion(question)
+                    chatViewModel.closeEditQuestionDialog()
+                }
             )
         }
     }
@@ -480,132 +484,6 @@ fun ChatInputBox(
             }
         )
     }
-}
-
-@Composable
-fun ChatTitleDialog(
-    initialTitle: String,
-    aiCoreModeEnabled: Boolean,
-    aiGeneratedResult: String,
-    isAICoreLoading: Boolean,
-    onDefaultTitleMode: () -> String?,
-    onAICoreTitleMode: () -> Unit,
-    onRetryRequest: () -> Unit,
-    onConfirmRequest: (title: String) -> Unit,
-    onDismissRequest: () -> Unit
-) {
-    val configuration = LocalConfiguration.current
-    var title by rememberSaveable { mutableStateOf(initialTitle) }
-    var useAICore by rememberSaveable { mutableStateOf(false) }
-    val untitledChat = stringResource(R.string.untitled_chat)
-
-    AlertDialog(
-        properties = DialogProperties(usePlatformDefaultWidth = false),
-        modifier = Modifier
-            .widthIn(max = configuration.screenWidthDp.dp - 40.dp)
-            .heightIn(max = configuration.screenHeightDp.dp - 80.dp),
-        title = { Text(text = stringResource(R.string.chat_title)) },
-        text = {
-            Column(modifier = Modifier.verticalScroll(rememberScrollState())) {
-                Text(text = stringResource(R.string.chat_title_dialog_description))
-                OutlinedTextField(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 8.dp, vertical = 16.dp),
-                    value = title,
-                    singleLine = true,
-                    isError = title.length > 50,
-                    supportingText = {
-                        if (title.length > 50) {
-                            Text(stringResource(R.string.title_length_limit, title.length))
-                        }
-                    },
-                    onValueChange = { title = it },
-                    label = { Text(stringResource(R.string.chat_title)) }
-                )
-                Row(
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    FilledTonalButton(
-                        modifier = Modifier
-                            .padding(horizontal = 8.dp)
-                            .height(48.dp)
-                            .weight(1F),
-                        enabled = !isAICoreLoading,
-                        onClick = { title = onDefaultTitleMode.invoke() ?: untitledChat }
-                    ) { Text(text = stringResource(R.string.default_mode)) }
-
-                    FilledTonalButton(
-                        enabled = aiCoreModeEnabled && !isAICoreLoading,
-                        modifier = Modifier
-                            .padding(horizontal = 8.dp)
-                            .height(48.dp)
-                            .weight(1F),
-                        onClick = {
-                            onAICoreTitleMode.invoke()
-                            useAICore = true
-                        }
-                    ) { Text(text = stringResource(R.string.ai_generated)) }
-                }
-
-                if (useAICore) {
-                    Card(
-                        colors = CardDefaults.cardColors(
-                            containerColor = MaterialTheme.colorScheme.surfaceVariant
-                        ),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 8.dp, vertical = 16.dp)
-                    ) {
-                        Column(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .heightIn(min = 64.dp)
-                                .padding(start = 20.dp, end = 20.dp, top = 20.dp, bottom = 8.dp)
-                        ) {
-                            Text(
-                                text = aiGeneratedResult.trimIndent() + if (isAICoreLoading) "▊" else "",
-                                fontWeight = FontWeight.Bold
-                            )
-                            Row(modifier = Modifier.fillMaxWidth()) {
-                                Spacer(Modifier.weight(1f))
-                                if (!isAICoreLoading) {
-                                    IconButton(
-                                        onClick = {
-                                            title = aiGeneratedResult.trimIndent().replace('\n', ' ')
-                                            useAICore = false
-                                        }
-                                    ) { Icon(Icons.Default.Done, contentDescription = stringResource(R.string.apply_generated_title)) }
-                                    IconButton(
-                                        onClick = onRetryRequest
-                                    ) { Icon(Icons.Rounded.Refresh, contentDescription = stringResource(R.string.retry_ai_title)) }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        onDismissRequest = onDismissRequest,
-        confirmButton = {
-            TextButton(
-                enabled = title.isNotBlank() && title != initialTitle,
-                onClick = {
-                    onConfirmRequest(title)
-                    onDismissRequest()
-                }
-            ) {
-                Text(stringResource(R.string.update))
-            }
-        },
-        dismissButton = {
-            TextButton(
-                onClick = onDismissRequest
-            ) {
-                Text(stringResource(R.string.cancel))
-            }
-        }
-    )
 }
 
 @Composable
